fixes issue #108 parsing announcements
